+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `IconBadge`: Add `border` and `backgroundColor` props style.
 - Fix: `ButtonFix`: Fix min width constraints
 
 ## [5.0.2] - 2021-10-13
@@ -48,7 +49,6 @@ Fixes:
 
 - `TagInput`: Remove console warning.
 - `TagInput`: Add accessible information for disabled items.
-
 
 ## [4.14.1] - 2021-10-08